+++ app/Http/Controllers/SystemSettingController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Controllers;
+
+class SystemSettingController extends Controller
+{
+    public function __construct()
+    {
+        $this->middleware('permission:settings.view');
+    }
+
+    public function index()
+    {
+        return view('settings.index');
+    }
+}